Maked #26 + #25 + #20
Fix crash cannot cast byte to integer caused by Crabzilla and SpikeZilla
Fix bugs from mob spawning
Fix a null pointer exception caused by UltimateFishHook
Fix a crash caused by java.lang.NullPointerException: Rendering entity in world(caused
y Twilight Fire ball)
Fix java.lang.Integer cannot be cast to java.lang.Byte caused by CrabZilla
fix flap_a1 + pincer_a1 + pincer_a2 + frog sound
Add a way to tp to the MLP Creature dimension (need rightclicking on AJInstance mob)?????????????????????????????

+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/CragadileInstance.java
@@ -27,7 +27,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
-// todo add a spawn for this entity
 public class CragadileInstance extends EntityMob {
 
     private GenericTargetSorterInstance TargetSorter;
@@ -495,15 +494,15 @@ public class CragadileInstance extends EntityMob {
     }
 
     public boolean getCanSpawnHere() {
-        for (int k = -6; k < 6; ++k) {
-            for (int j = -6; j < 6; ++j) {
-                for (int i = 0; i < 4; ++i) {
+        for (byte k = -6; k < 6; ++k) {
+            for (byte j = -6; j < 6; ++j) {
+                for (byte i = 0; i < 4; ++i) {
                     final Block bid = this.worldObj
-                        .getBlock((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
+                        .getBlock((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
                     if (bid == Blocks.mob_spawner) {
                         TileEntityMobSpawner tileentitymobspawner = null;
                         tileentitymobspawner = (TileEntityMobSpawner) this.worldObj
-                            .getTileEntity((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
+                            .getTileEntity((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
                         final String s = tileentitymobspawner.func_145881_a()
                             .getEntityNameToSpawn();
                         if (s != null && s.equals("MyCragadile")) {
