+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/FlameDragonInstance.java
@@ -34,7 +34,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
-// todo add a spawn for this entity
 public class FlameDragonInstance extends EntityTameable {
 
     private int boatPosRotationIncrements;
@@ -513,15 +512,15 @@ public class FlameDragonInstance extends EntityTameable {
     }
 
     public boolean getCanSpawnHere() {
-        for (int k = -5; k <= 5; ++k) {
-            for (int j = -5; j <= 5; ++j) {
-                for (int i = 3; i < 5; ++i) {
+        for (byte k = -5; k <= 5; ++k) {
+            for (byte j = -5; j <= 5; ++j) {
+                for (byte i = 3; i < 5; ++i) {
                     final Block bid = this.worldObj
-                        .getBlock((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
+                        .getBlock((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
                     if (bid == Blocks.mob_spawner) {
                         TileEntityMobSpawner tileentitymobspawner = null;
                         tileentitymobspawner = (TileEntityMobSpawner) this.worldObj
-                            .getTileEntity((int) this.posX + j, (int) this.posY + i, (int) this.posZ + k);
+                            .getTileEntity((byte) this.posX + j, (byte) this.posY + i, (byte) this.posZ + k);
                         final String s = tileentitymobspawner.func_145881_a()
                             .getEntityNameToSpawn();
                         if (s != null && s.equals("MyFlameDragon")) {
